requirements.txt
tqdm
numpy<2.0.0
opencv-python
pandas
#onnxruntime
onnxruntime-tools
Pillow
onnxruntime-gpu
matplotlib
seaborn
